fix(solver): reduce large conflicting-discriminant intersections to never (#14803)
`intersection_has_disjoint_object_literals` carried a hard
`MAX_DISJOINT_CHECK_SIZE = 25` cap that skipped the entire disjointness scan
once an intersection had more than 25 members, silently dropping the
never-reduction and producing a false-positive TS2322 for 26+ member
conflicting-discriminant chains (e.g. `{ x: 0 } & { x: 1 } & ... & { x: 25 }`).

Structural rule: when an object intersection shares a property NAME across two
members whose value-sets, primitive domains, or private/public visibility are
disjoint, tsc reduces the intersection to `never`; tsz now does this through
the solver's `intersection_has_disjoint_object_literals` regardless of member
count.

Replace the O(N^2) pairwise scan over whole shapes (and its size cap) with a
name-indexed per-name conflict test:
- index every member's properties by name once, O(sum of property counts);
- per name, run three existence tests that mirror the former per-pair
  predicate exactly — visibility collision, cross-domain literal-vs-primitive
  clash (honouring the both-optional exclusion), and literal value-set
  disjointness grouped by domain;
- the single-literal discriminant shape resolves in O(k), so the reduction is
  correct for arbitrarily large intersections with no quadratic cost and no
  silent drop.

Adjacent cases covered by new unit tests: number/string discriminants at 26 and
100 members, cross-domain literal-vs-primitive at scale, private/public
collisions at scale, and the negative guards (matching discriminants and
all-optional conflicts must NOT reduce to never).

Split the inline `mod tests` out of `normalize.rs` into `normalize_tests.rs`
(via `#[path]`) to keep the source under the 2000-line ceiling.
